Only one circuit break left; proper setup for namespace in acceptance tests
